Change what extra static assets are loaded
There are a couple of extra assets that we want to load on every single
page load, those are defcne.css (our sites CSS), the bootstrap.min.css
(Bootstrap's stock CSS) and then jQuery and bootstrap's javascript.

Do note that we load the javascript at the bottom, so that hopefully the
page will render more quickly and won't block while waiting for the
Javascript to download/execute.

This also renames the javascript block to javascript_head and adds a new
one called javascript_end. The idea being that we can select whether to
add a javascript to the head of the page or the end of the page. This
change also removes the block named "end", it is no longer required.