Move savepoint() into the while loop
With the savepoint() outside of the while loop, if anything were to go
wrong and we would rollback() there would be no future savepoint waiting
for us to happen, so rolling back in the future would not be possible.

Now we create it as we enter the while loop. Fixes that issue.