Add a HTTP forbidden view
If a user is not allowed access to a particular resource they are going
to get redirected to an HTTP forbidden view. If they are not logged in
they will get sent to the authentication page, if they are however
already logged in they simply don't have access and get denied access.