A new table for holding valid authentication tickets
The new UserTickets database holds the valid authentication tickets for
the current authentication ticket cookies that clients have. The idea
here is that we have a specific token for each user that is only valid
for a certain amount of time, AND can easily be revoked so that if you
log in from multiple computers you can log the other user out eventhough
the cookie is still valid.