Add configuration setting to add user to request
Using the current_user function in auth we now add an extra attribute to
each request as it comes in called .user. We reify it, meaning the
result gets cached after first run.

This allows us to take the database hit once instead of multiple times
when we are requesting the effective_principals or the
authenticated_userid.