Use traversal with URL dispatch
Using traversal we can make sure to set the appropriate ACL on an
object, and at the same time it will automatically pick up the right
view based on the name of the last element. This makes it much simpler
to wire up views based upon a URL, especially since we don't have to
duplicate a whole bunch of extra code.