If already logged in (has a valid authentication) redirect
When the user is already logged in we don't want them entering data into
forms they shouldn't. For example, if they are already logged in, there
should be no need for them to create a new user or validate a new user.
This simply re-directs those URL's to /user/ which is the users own
page.