If ?next=<URL> exists in auth, use it
If there is a query parameter named next that contains a path to the
next location the user should be sent then we use it to send the user
there.

This allows the forbidden view to send a user to the login page to
attempt to authenticate for the action they are trying to accomplish.