Change username/email to always be lowercase
I want to give the user the ability to log in using lower/uppercase of
their username, and also to make sure that no-one uses the same username
but with a different case.

Added a couple new hybrid properties that take care of it and where
necessary changed the function to make sure that username is lowercased
before going into the select.