Add bad_request error
If we get a POST request that doesn't match any single route/view then
we all through to the not found handlers, in this case we simply want to
return that it is a bad request, since POST generally has to go to a
defined endpoint.

This also gets returned if the CSRF check for instance doesn't match.