Implement password changing
If the user is logged in, and they are not logged in due to a password
reset procedure they get asked to enter their password. If they are
logged in due to a password reset request they get a form that allows
them to enter a password without the requirement for their previous
password.

Upon successfully setting a new password we set a flash message and
redirect them back to /user/.