More work is done in DefconEvent and Event ACL's
When we attempt to traverse to a DefconEvent we look and see if we can
find the short name in the database, if we can we return an Event()
object, passing it the found database information.

This allows us to dynamically create an __acl__ for it on the fly, which
can then be checked against the requesting user to see if they have
permission to do what they are asking to do.