Add new traversal acl node
This adds a new traversal node, called EventCreate, this way on the
add_view we can set the context to HTTPForbidden and we can set the
containment to include acl.EventCreate.

This means we get much finer grained control over what message is shown
when the user goes to a page where they may require authorization, but
we don't want to acknowledge that.